Move base type to separate file. Move both refactorings into single folder
